requirements/dev.txt
-r common.txt
django-browser-reload==1.6.0